+++ matthewplotlib/__init__.py
@@ -62,6 +62,7 @@ from matthewplotlib.plots import (
     hilbert,
     calendar,
     weeks,
+    table,
     text,
     border,
     axes,